xfs: widen xfs_refcount_irec fields to handle realtime refcountbt
Change the startblock and blockcount fields of xfs_refcount_irec to be
64 bits wide.  This enables us to use the same high level refcount code
for either tree.  We'll also collect all the resulting breakage fixes
here.

+++ fs/xfs/scrub/cow_repair.c
@@ -64,10 +64,10 @@ struct xrep_cow {
 	struct xfs_bmbt_irec	irec;
 
 	/* refcount btree block number of irec.br_startblock */
-	unsigned int		irec_startbno;
+	unsigned long long	irec_startbno;
 
 	/* refcount btree block number of the next refcount record we expect */
-	unsigned int		next_bno;
+	unsigned long long	next_bno;
 };
 
 /*
